Another certainly acceptable cast from an apr_off_t ... until we want to   adjust apr_sdbm entirely for largefile support
git-svn-id: https://svn.apache.org/repos/asf/apr/apr-util/trunk@58323 13f79535-47bb-0310-9956-ffa450edef68

+++ dbm/sdbm/sdbm_private.h
@@ -112,7 +112,7 @@ long sdbm_hash(const char *str, int len);
 #define SDBM_INVALIDATE_CACHE(db, finfo) \
     do { db->dirbno = (!finfo.size) ? 0 : -1; \
          db->pagbno = -1; \
-         db->maxbno = finfo.size * BYTESIZ; \
+         db->maxbno = (long)(finfo.size * BYTESIZ); \
     } while (0);
 
 #endif /* SDBM_PRIVATE_H */
